+++ src/modules/voxelrender/shaders/_shared.glsl
@@ -2,3 +2,5 @@
 $constant FlagOutline FLAGOUTLINE
 // only internal use
 #define FLAGHASNORMALPALETTECOLOR 2u
+// set on voxels that have the outline flag in vertex data (e.g. selected voxels)
+#define FLAGSELECTEDOUTLINE 4u